Version 4.16-BETA

+++ src/app/ui/records/search-records.component.ts
@@ -1,7 +1,7 @@
 import { HttpResponse } from '@angular/common/http';
 import { ChangeDetectionStrategy, Component, Injector, OnInit } from '@angular/core';
 import {
-  BaseRecordDataForSearch, BasicProfileFieldInput, CustomFieldDetailed, GeneralRecordsDataForSearch,
+  BaseRecordDataForSearch, BasicProfileFieldInput, CustomFieldDetailed, DeviceConfirmationTypeEnum, GeneralRecordsDataForSearch,
   GeneralRecordsQueryFilters, Group, RecordDataForSearch, RecordLayoutEnum,
   RecordQueryFilters, RecordResult, RecordWithOwnerResult, UserStatusEnum
 } from 'app/api/models';
@@ -53,41 +53,38 @@ export class SearchRecordsComponent
 
     // Get search data
     if (this.generalSearch) {
-      this.addSub(this.recordsService.getRecordDataForGeneralSearch({ type: this.type })
-        .subscribe(data => this.data = data));
+      this.addSub(
+        this.recordsService.getRecordDataForGeneralSearch({ type: this.type }).subscribe(data => this.data = data));
     } else {
       this.addSub(
-        this.recordsService.getRecordDataForOwnerSearch({ owner: this.param, type: this.type })
-          .subscribe(data => this.data = data));
+        this.recordsService.getRecordDataForOwnerSearch({ owner: this.param, type: this.type }).subscribe(data => this.data = data));
     }
   }
 
-  showKeywords(): boolean {
-    return !(this.data as RecordDataForSearch).hideKeywordsSearch || !this.isOwnRecords();
-  }
-
-  isOwnRecords(): boolean {
-    return !this.generalSearch && this.authHelper.isSelfOrOwner((this.data as RecordDataForSearch).user);
-  }
-
-  onDataInitialized(data: GeneralRecordsDataForSearch | RecordDataForSearch) {
-    if (!this.generalSearch && data.type.layout !== RecordLayoutEnum.LIST) {
-      throw new Error(`Invalid record layout: ${data.type.layout}`);
-    }
+  prepareForm(data: GeneralRecordsDataForSearch | RecordDataForSearch) {
     this.fieldsInSearch = data.customFields.filter(cf => data.fieldsInSearch.includes(cf.internalName));
-    this.fieldsInList = data.customFields.filter(cf => data.fieldsInList.includes(cf.internalName));
-    this.form.setControl('customValues', this.fieldHelper.customFieldsForSearchFormGroup(this.fieldsInSearch));
+    this.form.setControl('customValues', this.fieldHelper.customFieldsForSearchFormGroup(this.fieldsInSearch, data.query.customFields));
     if (this.generalSearch) {
       const general = data as GeneralRecordsDataForSearch;
       this.basicProfileFields = general.basicProfileFields;
       this.customProfileFields = general.customProfileFields;
-      this.groups = general.groups || [];
-      this.userStatuses = general.userStatuses || [];
       this.form.setControl('profileFields',
         this.fieldHelper.profileFieldsForSearchFormGroup(this.basicProfileFields, this.customProfileFields));
     }
-    this.form.patchValue(data.query);
+  }
 
+  onDataInitialized(data: GeneralRecordsDataForSearch | RecordDataForSearch) {
+    super.onDataInitialized(data);
+    if (!this.generalSearch && data.type.layout !== RecordLayoutEnum.LIST) {
+      throw new Error(`Invalid record layout: ${data.type.layout}`);
+    }
+    this.fieldsInList = data.customFields.filter(cf => data.fieldsInList.includes(cf.internalName));
+
+    if (this.generalSearch) {
+      const general = data as GeneralRecordsDataForSearch;
+      this.groups = general.groups || [];
+      this.userStatuses = general.userStatuses || [];
+    }
     const headingActions: HeadingAction[] = [];
     if (!this.generalSearch && data.create) {
       headingActions.push(new HeadingAction(SvgIcon.PlusCircle, this.i18n.general.addNew, () =>
@@ -106,28 +103,39 @@ export class SearchRecordsComponent
         }
       });
     this.headingActions = headingActions;
-    super.onDataInitialized(data);
+  }
+
+  showKeywords(): boolean {
+    return !(this.data as RecordDataForSearch).hideKeywordsSearch || !this.isOwnRecords();
+  }
+
+  isOwnRecords(): boolean {
+    return !this.generalSearch && this.authHelper.isSelfOrOwner((this.data as RecordDataForSearch).user);
   }
 
   protected doSearch(value: RecordSearchParams): Observable<HttpResponse<RecordResult[]>> {
-    return this.generalSearch ?
-      this.recordsService.searchGeneralRecords$Response(value) :
-      this.recordsService.searchOwnerRecords$Response(value);
+    return this.generalSearch ? this.recordsService.searchGeneralRecords$Response(value)
+      : this.recordsService.searchOwnerRecords$Response(value);
   }
 
   remove(record: RecordResult) {
-    this.confirmation.confirm({
-      message: this.i18n.general.removeItemConfirm,
-      callback: () => this.doRemove(record),
-    });
+    this.addSub(this.recordsService.getPasswordInputForRemoveRecord({ id: record.id }).subscribe(passwordInput =>
+      this.confirmation.confirm({
+        message: this.i18n.general.removeItemConfirm,
+        passwordInput,
+        createDeviceConfirmation: () => ({
+          type: DeviceConfirmationTypeEnum.MANAGE_RECORD,
+          recordType: this.data.type.id
+        }),
+        callback: params => this.doRemove(record, params.confirmationPassword),
+      })));
   }
 
-  private doRemove(record: RecordResult) {
-    this.addSub(this.recordsService.deleteRecord({ id: record.id })
-      .subscribe(() => {
-        this.notification.snackBar(this.i18n.general.removeItemDone);
-        this.update();
-      }));
+  private doRemove(record: RecordResult, confirmationPassword: string) {
+    this.addSub(this.recordsService.deleteRecord({ id: record.id, confirmationPassword }).subscribe(() => {
+      this.notification.snackBar(this.i18n.general.removeItemDone);
+      this.update();
+    }));
   }
 
   /**
